+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "create-nodets-app",
-  "version": "0.2.3",
+  "version": "0.2.4",
   "description": "My initial `NodeJs` project template using `TypeScript`.",
   "main": "./build",
   "scripts": {
@@ -42,7 +42,6 @@
     "@lib": "build/lib",
     "@interfaces": "build/interfaces",
     "@models": "build/models",
-    "@routes": "build/routes",
     "@schemas": "build/schemas"
   }
 }
