+++ src/commons/interfaces/example.ts
