+++ src/commons/interfaces/index.ts
@@ -1,0 +1,5 @@
+import iExample from './example'
+
+export {
+  iExample
+}
