+++ src/interfaces/grpc/client.ts
@@ -2,7 +2,7 @@ import * as grpc from '@grpc/grpc-js'
 import * as protoLoader from '@grpc/proto-loader'
 import { protoOption } from './protos'
 
-const protoPath = './src/grpc/protos/services.proto'
+const protoPath = './src/interfaces/grpc/protos/services.proto'
 
 const packageDefinition = protoLoader.loadSync(protoPath, protoOption)
 const { Example }: any = grpc.loadPackageDefinition(packageDefinition).services
@@ -15,22 +15,22 @@ const main = (port: number = 50051): void => {
     console.log('getMany', response)
   })
 
-  client.getOne({ params: { id: '62e2e05b3ea37c861d51815b' } }, (err: any, response: any): void => {
+  client.getOne({ params: { id: '62ced1a1f25c729e6e4c21d4' } }, (err: any, response: any): void => {
     if (err !== null && err !== undefined) throw new Error(err)
     console.log('getOne', response)
   })
 
-  client.postOne({ body: { name: 'Echo' } }, (err: any, response: any): void => {
+  client.postOne({ body: { name: 'Susi' } }, (err: any, response: any): void => {
     if (err !== null && err !== undefined) throw new Error(err)
     console.log('postOne', response)
   })
 
-  client.putOne({ params: { id: '62e2e05b3ea37c861d51815b' }, body: { name: 'Delta One' } }, (err: any, response: any): void => {
+  client.putOne({ params: { id: '62ced1a1f25c729e6e4c21d4' }, body: { name: 'Alfa Two' } }, (err: any, response: any): void => {
     if (err !== null && err !== undefined) throw new Error(err)
     console.log('putOne', response)
   })
 
-  client.delOne({ params: { id: '62e2e05b3ea37c861d51815b' } }, (err: any, response: any): void => {
+  client.delOne({ params: { id: '62ced1a1f25c729e6e4c21d4' } }, (err: any, response: any): void => {
     if (err !== null && err !== undefined) throw new Error(err)
     console.log('delOne', response)
   })
