+++ src/interfaces/grpc/protos/example.proto
