+++ src/interfaces/grpc/services/delOne/delOne.ts
