+++ src/interfaces/grpc/services/delOne/del_one.proto
