+++ src/interfaces/grpc/services/delOne/index.ts
