+++ src/interfaces/grpc/services/getMany/getMany.ts
