+++ src/interfaces/grpc/services/getMany/get_many.proto
