+++ src/interfaces/grpc/services/getMany/index.ts
