+++ src/interfaces/grpc/services/getOne/getOne.ts
