+++ src/interfaces/grpc/services/getOne/index.ts
