+++ src/interfaces/grpc/services/postOne/postOne.ts
