+++ src/interfaces/rest/index.ts
@@ -1,0 +1,9 @@
+import app from './app'
+
+export {
+  app
+}
+
+export default {
+  app
+}
