+++ src/interfaces/rest/routes/example/delOne.ts
