+++ src/interfaces/rest/routes/example/getMany.ts
