+++ src/interfaces/rest/routes/example/getOne.ts
