+++ src/interfaces/rest/routes/example/index.ts
