+++ src/interfaces/rest/routes/example/postOne.ts
