+++ src/interfaces/rest/routes/example/putOne.ts
